+++ src/main/resources/reference.conf
@@ -3,6 +3,20 @@
 # ==================================== #
 
 kamon {
+  agent {
+    show-banner = true
+    log-level = "INFO"
+
+    modules {
+      http4s-module {
+        stoppable = true
+        name = "http4s Instrumentation Module"
+        instrumentations = ["kamon.http4s.instrumentation.Http4sServerInstrumentation"]
+        within = ["org.http4s.server..*"]
+      }
+    }
+  }
+
   http4s {
     # Fully qualified name of the implementation of kamon.http4s.NameGenerator that will be used for assigning names
     # names to Spans.
