+++ src/main/scala/kamon/http4s/Metrics.scala
@@ -140,6 +140,4 @@ object Metrics {
   case class ServiceMetrics(generalMetrics: GeneralMetrics,
                             requestTimeMetrics: RequestTimeMetrics,
                             responseTimeMetrics: ResponseTimeMetrics)
-
-
 }
