+++ src/main/scala/kamon/http4s/instrumentation/package.scala
@@ -16,30 +16,35 @@
 
 package kamon.http4s
 
+import cats.effect.Sync
 import kamon.Kamon
 import kamon.context.{Context, TextMap}
 import org.http4s.{Header, Request}
+import cats.implicits._
+import org.slf4j.LoggerFactory
 
 package object instrumentation {
 
-  def decodeContext[F[_]](request: Request[F]): Context = {
-    val headersTextMap = readOnlyTextMapFromHeaders(request)
-    Kamon.contextCodec().HttpHeaders.decode(headersTextMap)
+  def decodeContext[F[_]:Sync](request: Request[F]): F[Context] = {
+    for {
+      headersTextMap <- readOnlyTextMapFromHeaders(request)
+      context <- Sync[F].delay(Kamon.contextCodec().HttpHeaders.decode(headersTextMap))
+    } yield context
   }
 
-  def encodeContext[F[_]](ctx:Context, request:Request[F]): Request[F] = {
+  def encodeContext[F[_]:Sync](ctx:Context, request:Request[F]): Request[F] = {
     val textMap = Kamon.contextCodec().HttpHeaders.encode(ctx)
     val headers = textMap.values.map{case (key, value) => Header(key, value)}
-    request//.putHeaders(headers.toSeq: _*)
+    request.putHeaders(headers.toSeq: _*)
   }
 
-  private def readOnlyTextMapFromHeaders[F[_]](request: Request[F]): TextMap = new TextMap {
+  def readOnlyTextMapFromHeaders[F[_]:Sync](request: Request[F]): F[TextMap] = Sync[F].delay(new TextMap {
     private val headersMap = request.headers.map(h => h.name.toString -> h.value).toMap
 
     override def values: Iterator[(String, String)] = headersMap.iterator
     override def get(key: String): Option[String] = headersMap.get(key)
     override def put(key: String, value: String): Unit = {}
-  }
+  })
 
   def isError(statusCode: Int): Boolean =
     statusCode >= 500 && statusCode < 600
@@ -47,4 +52,25 @@ package object instrumentation {
   object StatusCodes {
     val NotFound = 404
   }
+
+  trait Log[F[_]] {
+    def info(msg: String): F[Unit]
+    def warn(msg: String): F[Unit]
+    def debug(msg: String): F[Unit]
+    def error(error: Throwable): F[Unit]
+  }
+
+  object Log {
+    private val logger = LoggerFactory.getLogger(this.getClass)
+
+    implicit def syncLogInstance[F[_]](implicit F: Sync[F]): Log[F] =
+      new Log[F] {
+        override def info(msg: String): F[Unit] = F.delay(logger.info(msg))
+        override def warn(msg: String): F[Unit] = F.delay(logger.warn(msg))
+        override def debug(msg: String): F[Unit] = F.delay(logger.debug(msg))
+        override def error(error: Throwable): F[Unit] = F.delay(logger.error(error.getMessage, error))
+      }
+  }
 }
+
+
