+++ src/test/resources/logback.xml
@@ -5,7 +5,7 @@
         </encoder>
     </appender>
 
-    <root level="info">
+    <root level="ERROR">
         <appender-ref ref="STDOUT" />
     </root>
 </configuration>
